v2 phase 1c: extract shared Stage-2 structure resolution
resolveAccountStructure() in backfill/structure-stage.ts — both drivers
resolve attribution identically. Bootstrap keeps trust-on-hit; forward
gets the long-promised re-gather + fingerprint-compare path (classify
only on drift).
